rps_adk/requirements.txt
google-adk[extensions]>=0.1.0
      